+++ packages/sources/react/src/components/actions/VtmnDropdown/index.ts
@@ -1,0 +1,4 @@
+export { default as VtmnDropdown } from './VtmnDropdown';
+export * from './VtmnDropdown';
+export { default as VtmnDropdownItem } from './VtmnDropdownItem';
+export * from './VtmnDropdownItem';