+++ packages/sources/react/src/components/navigation/VtmnTabs/index.ts
@@ -1,0 +1,4 @@
+export { default as VtmnTabs } from './VtmnTabs';
+export * from './VtmnTabs';
+export { default as VtmnTabsItem } from './VtmnTabsItem';
+export * from './VtmnTabsItem';