+++ packages/sources/vue/src/components/actions/VtmnDropdown/index.ts
@@ -1,0 +1,8 @@
+import { App as Application } from 'vue';
+import VtmnDropdown from './VtmnDropdown.vue';
+
+VtmnDropdown.install = (Vue: Application) => {
+  Vue.component(VtmnDropdown.name, VtmnDropdown);
+};
+
+export default VtmnDropdown;