+++ packages/sources/vue/src/components/index.ts
@@ -1,28 +1,38 @@
 import VtmnButton from './actions/VtmnButton';
+import VtmnDropdown from './actions/VtmnDropdown';
+import VtmnDropdownItem from './actions/VtmnDropdownItem';
 import VtmnLink from './actions/VtmnLink';
+
 import VtmnTextInput from './forms/VtmnTextInput';
 import VtmnSelect from './forms/VtmnSelect';
+
 import VtmnBadge from './indicators/VtmnBadge';
 import VtmnLoader from './indicators/VtmnLoader';
 import VtmnPrice from './indicators/VtmnPrice';
 import VtmnProgressbar from './indicators/VtmnProgressbar';
 import VtmnTag from './indicators/VtmnTag';
 import VtmnRating from './indicators/VtmnRating';
+
 import VtmnBreadcrumb from './navigation/VtmnBreadcrumb';
 import VtmnBreadcrumbItem from './navigation/VtmnBreadcrumbItem';
 import VtmnNavbar from './navigation/VtmnNavbar';
 import VtmnSearch from './navigation/VtmnSearch';
+import VtmnTabs from './navigation/VtmnTabs';
+import VtmnTabsItem from './navigation/VtmnTabsItem';
+
 import VtmnAlert from './overlays/VtmnAlert';
 import VtmnModal from './overlays/VtmnModal';
 import VtmnPopover from './overlays/VtmnPopover';
 import VtmnSnackbar from './overlays/VtmnSnackbar';
 import VtmnToast from './overlays/VtmnToast';
 import VtmnTooltip from './overlays/VtmnTooltip';
+
 import VtmnCheckbox from './selection-controls/VtmnCheckbox';
 import VtmnChip from './selection-controls/VtmnChip';
 import VtmnQuantity from './selection-controls/VtmnQuantity';
 import VtmnRadioButton from './selection-controls/VtmnRadioButton';
 import VtmnToggle from './selection-controls/VtmnToggle';
+
 import VtmnAccordion from './structure/VtmnAccordion';
 import VtmnCard from './structure/VtmnCard';
 import VtmnDivider from './structure/VtmnDivider';
@@ -32,6 +42,8 @@ import VtmnSkeleton from './structure/VtmnSkeleton';
 
 export {
   VtmnButton,
+  VtmnDropdown,
+  VtmnDropdownItem,
   VtmnLink,
   VtmnTextInput,
   VtmnSelect,
@@ -45,6 +57,8 @@ export {
   VtmnBreadcrumbItem,
   VtmnNavbar,
   VtmnSearch,
+  VtmnTabs,
+  VtmnTabsItem,
   VtmnAlert,
   VtmnModal,
   VtmnPopover,