+++ packages/sources/vue/src/index.ts
@@ -14,6 +14,8 @@ export default install;
 
 // Actions
 export { default as VtmnButton } from './components/actions/VtmnButton';
+export { default as VtmnDropdown } from './components/actions/VtmnDropdown';
+export { default as VtmnDropdownItem } from './components/actions/VtmnDropdownItem';
 export { default as VtmnLink } from './components/actions/VtmnLink';
 // Forms
 export { default as VtmnSelect } from './components/forms/VtmnSelect';
@@ -30,6 +32,9 @@ export { default as VtmnBreadcrumb } from './components/navigation/VtmnBreadcrum
 export { default as VtmnBreadcrumbItem } from './components/navigation/VtmnBreadcrumbItem';
 export { default as VtmnNavbar } from './components/navigation/VtmnNavbar';
 export { default as VtmnSearch } from './components/navigation/VtmnSearch';
+export { default as VtmnTabs } from './components/navigation/VtmnTabs';
+export { default as VtmnTabsItem } from './components/navigation/VtmnTabsItem';
+
 // Overlays
 export { default as VtmnAlert } from './components/overlays/VtmnAlert';
 export { default as VtmnModal } from './components/overlays/VtmnModal';