+++ packages/sources/css/presets/tailwind/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.35.3](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.35.2...@vtmn/css-tailwind-preset@0.35.3) (2022-06-28)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
+## [0.35.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.35.1...@vtmn/css-tailwind-preset@0.35.2) (2022-06-28)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
+## [0.35.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.35.0...@vtmn/css-tailwind-preset@0.35.1) (2022-06-28)
+
+**Note:** Version bump only for package @vtmn/css-tailwind-preset
+
+
+
+
+
 # [0.35.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tailwind-preset@0.34.4...@vtmn/css-tailwind-preset@0.35.0) (2022-06-27)