+++ packages/sources/svelte/CHANGELOG.md
@@ -3,6 +3,69 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.49.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.49.1...@vtmn/svelte@0.49.2) (2022-06-28)
+
+**Note:** Version bump only for package @vtmn/svelte
+
+
+
+
+
+## [0.49.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.49.0...@vtmn/svelte@0.49.1) (2022-06-28)
+
+
+### Bug Fixes
+
+* **@vtmn/react, @vtmn/svelte:** sonarcloud code smells ([07a126d](https://github.com/Decathlon/vitamin-web/commit/07a126d38e81fadd0878cb62f83c8cdceeb7238d))
+
+
+
+
+
+# [0.49.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.48.1...@vtmn/svelte@0.49.0) (2022-06-28)
+
+
+### Features
+
+* **@vtmn/css, @vtmn/react, @vtmn/svelte, @vtmn/vue:** add or improve `VtmnTabs` component ([#1168](https://github.com/Decathlon/vitamin-web/issues/1168)) ([5c22816](https://github.com/Decathlon/vitamin-web/commit/5c22816a19a9e60e33080eec65bad02492b9dd39))
+
+
+
+
+
+## [0.48.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.48.0...@vtmn/svelte@0.48.1) (2022-06-28)
+
+
+### Bug Fixes
+
+* **svelte:** add correct types for Svelte popover component ([#1199](https://github.com/Decathlon/vitamin-web/issues/1199)) ([e510ac8](https://github.com/Decathlon/vitamin-web/commit/e510ac8241a23d907ded71087598b5c8cf5090c4))
+
+
+
+
+
+# [0.48.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.47.0...@vtmn/svelte@0.48.0) (2022-06-28)
+
+
+### Features
+
+* **@vtmn/css, @vtmn/react, @vtmn/svelte:** `progressbar`, add a11y requirements ([#1129](https://github.com/Decathlon/vitamin-web/issues/1129)) ([69829e8](https://github.com/Decathlon/vitamin-web/commit/69829e88f2cdc5f80b242cc4c5c2819d8c83d610))
+
+
+
+
+
+# [0.47.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.46.5...@vtmn/svelte@0.47.0) (2022-06-28)
+
+
+### Features
+
+* **@vtmn/css, @vtmn/react, @vtmn/svelte, @vtmn/vue:** `divider`, add a11y requirements ([#1137](https://github.com/Decathlon/vitamin-web/issues/1137)) ([d651cd2](https://github.com/Decathlon/vitamin-web/commit/d651cd2afbd5b4381865d04e5c5c27f671812ae8))
+
+
+
+
+
 ## [0.46.5](https://github.com/Decathlon/vitamin-web/compare/@vtmn/svelte@0.46.4...@vtmn/svelte@0.46.5) (2022-06-27)
 
 **Note:** Version bump only for package @vtmn/svelte