+++ packages/sources/vue/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.41.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.41.1...@vtmn/vue@0.41.2) (2022-06-28)
+
+
+### Bug Fixes
+
+* **@vtmn/css:** sonarcloud code smells ([f95800e](https://github.com/Decathlon/vitamin-web/commit/f95800e71b58cd015ad6bf6f0ad64283f2193df0))
+
+
+
+
+
+## [0.41.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.41.0...@vtmn/vue@0.41.1) (2022-06-28)
+
+
+### Bug Fixes
+
+* **@vtmn/vue:** sonarcloud code smells ([e9dfc77](https://github.com/Decathlon/vitamin-web/commit/e9dfc77d9861544c1edbcdcefe02de96a1aba568))
+
+
+
+
+
+# [0.41.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.40.0...@vtmn/vue@0.41.0) (2022-06-28)
+
+
+### Features
+
+* **@vtmn/css, @vtmn/react, @vtmn/svelte, @vtmn/vue:** add or improve `VtmnTabs` component ([#1168](https://github.com/Decathlon/vitamin-web/issues/1168)) ([5c22816](https://github.com/Decathlon/vitamin-web/commit/5c22816a19a9e60e33080eec65bad02492b9dd39))
+* **@vtmn/react, @vtmn/vue:** add `VtmnDropdown` component ([#1169](https://github.com/Decathlon/vitamin-web/issues/1169)) ([4e88194](https://github.com/Decathlon/vitamin-web/commit/4e881948598a7c0acda21d89f3db03d39becaeb6))
+
+
+
+
+
+# [0.40.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.39.0...@vtmn/vue@0.40.0) (2022-06-28)
+
+
+### Features
+
+* **@vtmn/css, @vtmn/react, @vtmn/svelte:** `progressbar`, add a11y requirements ([#1129](https://github.com/Decathlon/vitamin-web/issues/1129)) ([69829e8](https://github.com/Decathlon/vitamin-web/commit/69829e88f2cdc5f80b242cc4c5c2819d8c83d610))
+
+
+
+
+
+# [0.39.0](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.38.5...@vtmn/vue@0.39.0) (2022-06-28)
+
+
+### Features
+
+* **@vtmn/css, @vtmn/react, @vtmn/svelte, @vtmn/vue:** `divider`, add a11y requirements ([#1137](https://github.com/Decathlon/vitamin-web/issues/1137)) ([d651cd2](https://github.com/Decathlon/vitamin-web/commit/d651cd2afbd5b4381865d04e5c5c27f671812ae8))
+
+
+
+
+
 ## [0.38.5](https://github.com/Decathlon/vitamin-web/compare/@vtmn/vue@0.38.4...@vtmn/vue@0.38.5) (2022-06-27)
 
 **Note:** Version bump only for package @vtmn/vue